The First Step
This is the first Stable Master We have. Clone it And mess around.
This is up to the "Cammera Follows Player" Video.

I made this neat and tidy. Every script group has a folder, every material group has a folder. If possible, try to keep it that way. :)

Good Luck!